% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_design.R
\name{baGetDesign}
\alias{baGetDesign}
\title{Get designs}
\usage{
baGetDesign(par.design, lst.setting = NULL, design = c("simon",
  "truth", "betabinom"), f.simu = baSimuTcell, rho = NULL,
  adj.rho = 1, bsize = 1, nmin = 10, nmax = 100, nreps = 10000,
  ...)
}
\description{
Get designs
}
