% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_simu.R
\name{baSimuBatchError}
\alias{baSimuBatchError}
\title{Simulate errors for all batches}
\usage{
baSimuBatchError(batch.sizes, par.err = list(gamma = list(error.type =
  "normal", sig = 1), delta = list(error.type = "normal", sig = 1), epsilon
  = list(error.type = "normal", sig = 1)))
}
\description{
Simulate errors for all batches
}
