% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basim_simu.R
\name{baSimuBetaBin}
\alias{baSimuBetaBin}
\title{Simulate batch data from Beta-Binomial}
\usage{
baSimuBetaBin(batch.sizes, p, rho = 0, seed = NULL)
}
\arguments{
\item{batch.sizes}{number of subjects in each batch}

\item{p}{mean of the Beta distribution a/(a+b)}

\item{rho}{inter-class correlation 1/(a+b+1)}
}
\value{
Simulated binary outcomes by batches
}
\description{
Simulate batch data from Beta-Binomial
}
